+++ mahasiswa/mahasiswaDelete.php
@@ -1,6 +1,13 @@
 <?php
 include "../koneksi/koneksi.php";
 
+session_start();
+
+if (!isset($_SESSION['role']) || ($_SESSION['role'] !== 'admin')) {
+  header("Location: ../index.php");
+  exit();
+}
+
 $nim = $_GET["nim"];
 $delMhs = "DELETE FROM MAHASISWA WHERE nim='$nim'";
 $resultMhs = mysqli_query($koneksi, $delMhs);
